requirements.txt
pyspark==3.5.0
requests==2.31.0
pyyaml==6.0.1
matplotlib==3.8.2
pandas==2.1.4
